one approach to keeping test pgks up to date
we need the versions of core packages installed via the test poetry.lock up to date with the versions of what's actually installed locally, otherwise dask may not be able to connect, etc.

this way just uses the base poetry.lock instead of a different fixture for tests. to try to save time during tests, we set the PIP_PACKAGES env var from pytest->docker-compose->docker build, to build the docker image wth packages already installed (so they aren't re-installed every time the test runs).

Problem is that the bloated coiled package is a dep of sneks, and pip decides to solve for different transitive dependencies of it than poetry does.

Also this might be tricky when using PDM.